tiny fix

+++ problems/4/shared.h
@@ -5,5 +5,5 @@
 
 // initialized by hakase.cc (before friend app is executed)
 extern int64_t SHARED_SYMBOL(sync_flag);
-extern uint32_t SHARED_SYMBOL(test_value[0x20]);
+extern uint64_t SHARED_SYMBOL(test_value[0x20]);
 extern void* SHARED_SYMBOL(hpet_addr);
